@@ -1,18 +1,19 @@
-import React from 'react';
-import styles from './Chat.pcss';
-import ChatMessages from './ChatMessages';
+import React from "react";
+import styles from "./Chat.pcss";
+import ChatMessages from "./ChatMessages";
+import PropTypes from "prop-types";
 
 export default function Chat({ yali, sendMessage, messages }) {
   return (
     <div className={styles.root}>
-      <h2>Latteudet</h2>
+      <h2>Tsätti</h2>
       <ChatMessages messages={messages} />
     </div>
   );
 }
 
 Chat.propTypes = {
-  yali: React.PropTypes.object.isRequired,
-  sendMessage: React.PropTypes.func.isRequired,
-  messages: React.PropTypes.object.isRequired,
+  yali: PropTypes.object.isRequired,
+  sendMessage: PropTypes.func.isRequired,
+  messages: PropTypes.object.isRequired
 };
