@@ -1,20 +1,22 @@
-import React from 'react';
-import styles from './Consumables.pcss';
-import Consumable from './Consumable';
-import consumables from 'services/consumables';
+import React from "react";
+import styles from "./Consumables.pcss";
+import Consumable from "./Consumable";
+import consumables from "services/consumables";
 
 export default function Consumables(props) {
-
   const { consume } = props;
 
   return (
     <section className={styles.root}>
-
-      {consumables.toList().map(consumable =>
-        <Consumable consume={consume} key={consumable.type} consumable={consumable} />
-      )}
-
+      {consumables
+        .toList()
+        .map(consumable => (
+          <Consumable
+            consume={consume}
+            key={consumable.type}
+            consumable={consumable}
+          />
+        ))}
     </section>
   );
-
 }
