@@ -1,14 +1,17 @@
-import React from 'react';
-import styles from './Yali.pcss';
-import Mouth from './yali/Mouth';
-import yaliBg from 'images/yali.png';
-import { ItemTypes } from 'constants';
-import { DropTarget } from 'react-dnd';
+import React from "react";
+import styles from "./Yali.pcss";
+import Mouth from "./yali/Mouth";
+import yaliBg from "../assets/images/yali.png";
+import { ItemTypes } from "../constants";
+import { DropTarget } from "react-dnd";
 
 function Yali({ yali, isOver, connectDropTarget, consumable }) {
   return connectDropTarget(
-    <section className={styles.root} style={{ backgroundImage: `url('${yaliBg}')` }}>
-      <Mouth mood={yali.get('mood')} />
+    <section
+      className={styles.root}
+      style={{ backgroundImage: `url('${yaliBg}')` }}
+    >
+      <Mouth mood={yali.get("mood")} />
     </section>
   );
 }
@@ -23,7 +26,6 @@ export default DropTarget(
     }
   },
   (connect, monitor) => {
-
     return {
       connectDropTarget: connect.dropTarget(),
       isOver: monitor.isOver(),
