@@ -1,42 +1,39 @@
-import { Map } from 'immutable';
-import { calculateMood } from 'services/yali';
+import { Map } from "immutable";
+import { calculateMood } from "services/yali";
 
-export const TICK = 'TICK';
-export const CONSUME = 'CONSUME';
+export const TICK = "TICK";
+export const CONSUME = "CONSUME";
 
 export function tick() {
   return {
-    type: TICK,
+    type: TICK
   };
 }
 
 export function consume(consumable) {
   return {
     type: CONSUME,
-    payload: consumable,
+    payload: consumable
   };
 }
 
 const defaultState = Map({
   alcohol: 90,
-  fullness: 90,
-}).update(state => state.set('mood', calculateMood(state)));
+  fullness: 90
+}).update(state => state.set("mood", calculateMood(state)));
 
-export default function (state = defaultState, action) {
+export default function(state = defaultState, action) {
   switch (action.type) {
     case TICK:
       return state
-        .update('fullness', f => (
-          (f <= 0) ? 0 : f - 0.1
-        ))
-        .update('alcohol', a => (
-          (a <= 0) ? 0 : a - 0.25
-        ))
-        .update(s => s.set('mood', calculateMood(s)));
+        .update("fullness", f => (f <= 0 ? 0 : f - 0.1))
+        .update("alcohol", a => (a <= 0 ? 0 : a - 0.25))
+        .update(s => s.set("mood", calculateMood(s)));
 
     case CONSUME:
-      return action.payload.consume(state)
-        .update(s => s.set('mood', calculateMood(s)));
+      return action.payload
+        .consume(state)
+        .update(s => s.set("mood", calculateMood(s)));
     default:
       return state;
   }
